add pytest dispatcher for MPI test in test_packages
